Resolve CHANGELOG ordering between Phase 1D-1 REST flags and Phase 1D-2 long-read topic after merging main.

+++ core/data/src/main/kotlin/fr/forumhfr/redface2/core/data/forum/RestForumDtos.kt
@@ -68,7 +68,7 @@ internal data class RestTopic(
     @SerialName("is_read") val isRead: Boolean? = null,
     @SerialName("flag_owntopic") val flagOwntopic: Int? = null,
     @SerialName("last_position") val lastPosition: Int? = null,
-    @SerialName("last_post_read_id") val lastPostReadId: Int? = null,
+    @SerialName("last_post_read_id") val lastPostReadId: Long? = null,
 )
 
 @Serializable
